django_ml/requirements.txt
absl-py==0.11.0
argon2-cffi==20.1.0
asgiref==3.2.7
astunparse==1.6.3
autopep8==1.5.4
bcrypt==3.1.7
boto3==1.14.63
botocore==1.17.63
cached-property==1.5.2
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
Django==3.0.3
django-cors-headers==3.5.0
djangorestframework==3.12.1
docutils==0.15.2
Faker==4.1.1
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
jmespath==0.10.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
Markdown==3.3.3
numpy==1.18.5
oauthlib==3.1.0
opencv-python-headless==4.4.0.46
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
scipy==1.5.3
six==1.15.0
sqlparse==0.3.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-cpu==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.1
urllib3==1.25.10
Werkzeug==1.0.1
whitenoise==5.2.0
wincertstore==0.2
wrapt==1.12.1
